+++ core/search/contacts_test.go
@@ -123,7 +123,7 @@ func TestDeindexContacts(t *testing.T) {
 
 	// ensures changes are visible in elastic
 	refreshElastic := func() {
-		_, err := rt.ES.Indices.Refresh().Index(rt.Config.ElasticContactsIndex).Do(ctx)
+		_, err := rt.ES.Client.Indices.Refresh().Index(rt.Config.ElasticContactsIndex).Do(ctx)
 		require.NoError(t, err)
 	}
 
@@ -132,16 +132,17 @@ func TestDeindexContacts(t *testing.T) {
 	assertSearchCount(t, rt, elastic.Term("org_id", testdb.Org1.ID), 124)
 	assertSearchCount(t, rt, elastic.Term("org_id", testdb.Org2.ID), 121)
 
-	deindexed, err := search.DeindexContactsByID(ctx, rt, testdb.Org1.ID, []models.ContactID{testdb.Bob.ID, testdb.Cat.ID})
+	esDeindexed, osDeindexed, err := search.DeindexContactsByID(ctx, rt, testdb.Org1.ID, []models.ContactID{testdb.Bob.ID, testdb.Cat.ID})
 	assert.NoError(t, err)
-	assert.Equal(t, 2, deindexed)
+	assert.Equal(t, 2, esDeindexed)
+	assert.Equal(t, 0, osDeindexed)
 
 	refreshElastic()
 
 	assertSearchCount(t, rt, elastic.Term("org_id", testdb.Org1.ID), 122)
 	assertSearchCount(t, rt, elastic.Term("org_id", testdb.Org2.ID), 121)
 
-	deindexed, err = search.DeindexContactsByOrg(ctx, rt, testdb.Org1.ID, 100)
+	deindexed, err := search.DeindexContactsByOrg(ctx, rt, testdb.Org1.ID, 100)
 	assert.NoError(t, err)
 	assert.Equal(t, 100, deindexed)
 
@@ -167,7 +168,7 @@ func TestDeindexContacts(t *testing.T) {
 func assertSearchCount(t *testing.T, rt *runtime.Runtime, query elastic.Query, expected int) {
 	src := map[string]any{"query": query}
 
-	resp, err := rt.ES.Count().Index(rt.Config.ElasticContactsIndex).Raw(bytes.NewReader(jsonx.MustMarshal(src))).Do(context.Background())
+	resp, err := rt.ES.Client.Count().Index(rt.Config.ElasticContactsIndex).Raw(bytes.NewReader(jsonx.MustMarshal(src))).Do(context.Background())
 	require.NoError(t, err)
 	assert.Equal(t, expected, int(resp.Count))
 }